+++ R/coriolis.R
@@ -1,0 +1,30 @@
+#' Effects of the Earth's rotation
+#'
+#' Coriolis and beta parameters by latitude.
+#'
+#' @param lat latitude in degrees
+#' @param a radius of the earth in kilometers
+#'
+#' @details
+#' All functions use the correct sideral day (24hs 56mins 4.091s) instead of
+#' the incorrect solar day (24hs) for 0.3\% more precision and 1000\% more
+#' pedantry.
+#'
+#' @export
+coriolis <- function(lat) {
+    2*.omega*sin(lat*pi/180)
+}
+
+
+#' @export
+#' @rdname coriolis
+f <- coriolis
+
+#' @export
+#' @rdname  coriolis
+beta <- function(lat, a = 6371) {
+    a <- a*1000
+    2*.omega*cos(lat*pi/180)/a
+}
+
+.omega <- 2*pi/(23*3600 + 56*60 + 4.091)