+++ NEWS.md
@@ -1,7 +1,9 @@
 # metR 0.0.9101
-- New function: GetTopography()
+- New function: `GetTopography()`
 - ReadNetCDF() now can output arrays and vectors.
 - Changed name. 
+- New website ([https://eliocamp.github.io/metR/](https://eliocamp.github.io/metR/)) with documentation. Thanks to `pkgdown`.
+- New functions: `coriolis()` and `beta()`.
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.